fix: icon not shown for negative `tag` and `badge`, wrong warning icon for Orange theme (Orange-OpenSource/ouds-ios#1216) (Orange-OpenSource/ouds-ios#1219) (#165)
Related to Orange-OpenSource/ouds-ios#1216
Related to Orange-OpenSource/ouds-ios#1219

+++ DesignToolbox/DesignToolbox/Pages/Components/Indicators/Badge/BadgeConfigurationView.swift
@@ -53,8 +53,8 @@ final class BadgeConfigurationModel: ComponentConfiguration {
 
     var themeName: String
 
-    var count: UInt {
-        UInt(countText) ?? 1
+    var count: UInt8 {
+        UInt8(countText) ?? 1
     }
 
     var enableFlipIcon: Bool {
